+++ app/src/main/java/com/reisdeveloper/githubapiconsumer/di/ui/MainActivity.kt
@@ -1,15 +1,12 @@
-package com.reisdeveloper.githubapiconsumer.ui
+package com.reisdeveloper.githubapiconsumer.di.ui
 
 import android.os.Bundle
-import com.google.android.material.snackbar.Snackbar
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.WindowCompat
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupActionBarWithNavController
-import android.view.Menu
-import android.view.MenuItem
 import com.reisdeveloper.githubapiconsumer.R
 import com.reisdeveloper.githubapiconsumer.databinding.ActivityMainBinding
