+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".ui.home.HomeFragment">
+    tools:context=".di.ui.home.HomeFragment">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/textInputSearchUser"
