+++ app/src/main/res/layout/fragment_user_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.user.UserFragment">
+    tools:context=".di.ui.user.UserFragment">
 
     <ImageView
         android:id="@+id/imgUser"
