+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/HomeFragment"
-        android:name="com.reisdeveloper.githubapiconsumer.ui.home.HomeFragment"
+        android:name="com.reisdeveloper.githubapiconsumer.di.ui.home.HomeFragment"
         android:label="@string/home_fragment_label"
         tools:layout="@layout/fragment_home">
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.reisdeveloper.githubapiconsumer.ui.user.UserFragment"
+        android:name="com.reisdeveloper.githubapiconsumer.di.ui.user.UserFragment"
         android:label="@string/user_fragment_label"
         tools:layout="@layout/fragment_user_detail">
     </fragment>
